feat(path): add support of URL input for basename, dirname, extname, join, and normalize
