+++ path/normalize_test.ts
@@ -2,8 +2,6 @@
 import { assertEquals } from "@std/assert";
 import * as windows from "./windows/mod.ts";
 import * as posix from "./posix/mod.ts";
-import { normalize as windowsUnstableNormalize } from "./windows/unstable_normalize.ts";
-import { normalize as posixUnstableNormalize } from "./posix/unstable_normalize.ts";
 
 Deno.test(`normalize() returns "." if input is empty`, function () {
   assertEquals(posix.normalize(""), ".");
@@ -16,7 +14,7 @@ Deno.test("posix.normalize() normalizes posix specific paths", () => {
     "/foo/bar/baz/asdf",
   );
   assertEquals(
-    posixUnstableNormalize(new URL("file:///foo/bar//baz/asdf/quux/..")),
+    posix.normalize(new URL("file:///foo/bar//baz/asdf/quux/..")),
     "/foo/bar/baz/asdf/",
   );
 });
@@ -27,7 +25,7 @@ Deno.test("windows.normalize() normalizes windows specific paths", () => {
     "\\\\server\\share\\dir\\file.ext",
   );
   assertEquals(
-    windowsUnstableNormalize(new URL("file:///C:/foo/bar/../baz/quux")),
+    windows.normalize(new URL("file:///C:/foo/bar/../baz/quux")),
     "C:\\foo\\baz\\quux",
   );
 });
