+++ path/posix/join.ts
@@ -2,6 +2,7 @@
 // This module is browser compatible.
 
 import { assertPath } from "../_common/assert_path.ts";
+import { fromFileUrl } from "./from_file_url.ts";
 import { normalize } from "./normalize.ts";
 
 /**
@@ -12,8 +13,8 @@ import { normalize } from "./normalize.ts";
  * import { join } from "@std/path/posix/join";
  * import { assertEquals } from "@std/assert";
  *
- * const path = join("/foo", "bar", "baz/asdf", "quux", "..");
- * assertEquals(path, "/foo/bar/baz/asdf");
+ * assertEquals(join("/foo", "bar", "baz/asdf", "quux", ".."), "/foo/bar/baz/asdf");
+ * assertEquals(join(new URL("file:///foo"), "bar", "baz/asdf", "quux", ".."), "/foo/bar/baz/asdf");
  * ```
  *
  * @example Working with URLs
@@ -29,14 +30,16 @@ import { normalize } from "./normalize.ts";
  * assertEquals(url.href, "https://deno.land/std/path/mod.ts");
  * ```
  *
- * Note: If you are working with file URLs,
- * use the new version of `join` from `@std/path/posix/unstable-join`.
- *
+ * @param path
  * @param paths The paths to join.
  * @returns The joined path.
  */
-export function join(...paths: string[]): string {
-  if (paths.length === 0) return ".";
+export function join(path?: URL | string, ...paths: string[]): string {
+  if (path === undefined) return ".";
+  if (path instanceof URL) {
+    path = fromFileUrl(path);
+  }
+  paths = path ? [path, ...paths] : paths;
   paths.forEach((path) => assertPath(path));
   const joined = paths.filter((path) => path.length > 0).join("/");
   return joined === "" ? "." : normalize(joined);
