+++ path/windows/extname.ts
@@ -4,6 +4,7 @@
 import { CHAR_COLON, CHAR_DOT } from "../_common/constants.ts";
 import { assertPath } from "../_common/assert_path.ts";
 import { isPathSeparator, isWindowsDeviceRoot } from "./_util.ts";
+import { fromFileUrl } from "./from_file_url.ts";
 
 /**
  * Return the extension of the `path` with leading period.
@@ -13,17 +14,17 @@ import { isPathSeparator, isWindowsDeviceRoot } from "./_util.ts";
  * import { extname } from "@std/path/windows/extname";
  * import { assertEquals } from "@std/assert";
  *
- * const ext = extname("file.ts");
- * assertEquals(ext, ".ts");
+ * assertEquals(extname("file.ts"), ".ts");
+ * assertEquals(extname(new URL("file:///C:/foo/bar/baz.ext")), ".ext");
  * ```
- *
- * Note: If you are working with file URLs,
- * use the new version of `extname` from `@std/path/windows/unstable-extname`.
  *
  * @param path The path to get the extension from.
  * @returns The extension of the `path`.
  */
-export function extname(path: string): string {
+export function extname(path: string | URL): string {
+  if (path instanceof URL) {
+    path = fromFileUrl(path);
+  }
   assertPath(path);
 
   let start = 0;
