+++ path/windows/join.ts
@@ -4,6 +4,7 @@
 import { assertPath } from "../_common/assert_path.ts";
 import { isPathSeparator } from "./_util.ts";
 import { normalize } from "./normalize.ts";
+import { fromFileUrl } from "./from_file_url.ts";
 
 /**
  * Join all given a sequence of `paths`,then normalizes the resulting path.
@@ -13,17 +14,19 @@ import { normalize } from "./normalize.ts";
  * import { join } from "@std/path/windows/join";
  * import { assertEquals } from "@std/assert";
  *
- * const joined = join("C:\\foo", "bar", "baz\\..");
- * assertEquals(joined, "C:\\foo\\bar");
+ * assertEquals(join("C:\\foo", "bar", "baz\\.."), "C:\\foo\\bar");
+ * assertEquals(join(new URL("file:///C:/foo"), "bar", "baz\\.."), "C:\\foo\\bar");
  * ```
  *
- * Note: If you are working with file URLs,
- * use the new version of `join` from `@std/path/windows/unstable-join`.
- *
+ * @param path The path to join. This can be string or file URL.
  * @param paths The paths to join.
  * @returns The joined path.
  */
-export function join(...paths: string[]): string {
+export function join(path?: URL | string, ...paths: string[]): string {
+  if (path instanceof URL) {
+    path = fromFileUrl(path);
+  }
+  paths = path ? [path, ...paths] : paths;
   paths.forEach((path) => assertPath(path));
   paths = paths.filter((path) => path.length > 0);
   if (paths.length === 0) return ".";
